backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.3.1
python-dotenv==1.0.0
pypdf2==3.0.1
python-docx==0.8.11
openai==1.3.5
boto3==1.29.3
redis==5.0.1
pytest==7.4.3
httpx==0.25.2
spacy==3.7.2
